requirements.txt
agent-os-kernel[full]
agentmesh-runtime
inter-agent-trust-protocol
fastapi
uvicorn